+++ app/src/main/java/org/mozilla/fenix/home/sessioncontrol/SessionControlComponent.kt
@@ -80,7 +80,6 @@ typealias TabCollection = ACTabCollection
 sealed class CollectionAction : Action {
     data class Expand(val collection: TabCollection) : CollectionAction()
     data class Collapse(val collection: TabCollection) : CollectionAction()
-    data class OpenTabs(val collection: TabCollection) : CollectionAction()
 }
 
 sealed class SessionControlAction : Action {
